Fixes a decoding error when data arrives in multiple parts. Adds a precondition to prevent int parsing sizes for values too large for the platform.
Motivation:
To prevent errors when using on networks or platforms of 32bit or less.

Modifications:
Added a decoder reading state to track if reading the header or body of the data.
When the reading state reads the header it stores then length of the body.
Added a precondition in the constructor to check the set byte width against the platform.

Result:
The decoding should work when data is delayed.
This should prevent a DOS attack by reading arbitrary amount of data on lower bit platforms.