Moves the ByteLength structure to be inside the decoder class. Improves the split data unit test and removes a potential bug with the splitting of data.
Motivation:
To remove the potential for error with the splitting of incoming data. Also, to neaten the code.

Modifications:
Moved the ‘ByteLength’ structure definition.
Moved a public method definition away for 3 related methods.
Improved the split data unit test to split into single bytes.
Removed a double read of the body data which may have caused bugs.

Result:
Less potential for error.
Improved tests.
Neater code that is a little easier to follow.