Ensures that the LengthFieldPrepender class reports a sensible error if the message is too long for the header.
Motivation:
To enhance error reporting.

Modifications:
Added a LengthFieldPrependerError which has a messageDataTooLongForLengthField that is thrown if the length is too long.
An additional test to check that the error is thrown.
Some neatening of comments.

Result:
A clear error will be returned is the length is too long for the header.