Handle absolute-form requests when converting from/to HTTPTypes (#293)
Handle absolute-form requests when converting from/to HTTPTypes.

### Motivation:

HTTPTypes represent modern HTTP/2 / HTTP/3 semantics which do not
support absolute-form requests.

### Modifications:

When converting an absolute-form request to an HTTPRequest, the request
target is converted to its path. When converting an HTTPRequest back,
the user needs to specify which form to convert to.

### Result:

HTTPRequest no longer contains invalid paths when working with cleartext
HTTP/1 proxies.